requirements.txt
streamlit>=1.22.0
pandas>=1.5.3
numpy>=1.24.3
torch>=2.0.0
networkx>=3.1
matplotlib>=3.7.1
Pillow>=9.5.0
PyPDF2>=3.0.0 